ref!: remove old rts from ui and delegate to the new rts engine
